requirements.txt
matplotlib
numpy
torch
torchvision
pandas
imageio
scikit-learn
Pillow
scipy
sklearn
pyserial
